requirements.txt
# OpsPilot Dependencies
# Python 3.9+ recommended

# Excel 解析
openpyxl>=3.1.0
pandas>=2.0.0

# Word 文档生成
python-docx>=1.0.0
docxtpl>=0.17.0

# YAML 配置解析
pyyaml>=6.0

# 命令行参数
click>=8.1.0

# MCP 服务化
fastmcp>=0.1.0

# Web 配置中心
flask>=3.0.0
werkzeug>=3.0.0

# LLM 客户端
requests>=2.28.0
openai>=1.0.0
anthropic>=0.18.0

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0
